Fix case when plural form of root names clash for plural names classes
If the class name is plural - eg. Series - the generated root names with the
plural form may generate the same vaslue for select/selectByPk, insert/insertOne
etc. To resolve this, if names clash we add -es suffix to the plural operations
and also print a warning about it a suggest to use @HasuraRootNames to
have better names.