Fix adding extra table names which are neither Java entities, nor join tables
Some joined types, like

    @ElementCollection
    private val names: List<String>? = null

are neither many-to-many joins, nor they join a Java entity. In this case
we handle these join tables as extra table names, which need to be tracked
so that Hasura array relation definition can be applied, but will have
no other configuration on such tables (for now).