Fix case of recursive many-to-many tables
If the many-to-many join table refers on both ends to the same table the
generated object relationship names would be the same for both ends (as the
name is derived form the related table's name) and so it would cause clash.

To avoid this the clashing object relationship will be added a "2" postfix and
a warning is displayed suggesting to use @HasuraAlias(..., joinFieldAlias).