Get rid of old code add metadataFile and loadMetadata support
Old configuration code has been removed.

Similar to confFile and loadConf added metadataFile and loadMetadata.
configure() will save metadata JSON to metadataFile and if loadMetadata is true,
will load the metadata to Hasura.

From now on loadMetadata is the preferred way to configure Hasura as it is
way faster the the 'bulk' generated in confFile.