Add support for loading cascaded delete definitions separately
When using the new metadata format for configuring Hasura, this format doesn't
include the cascaded delete sql definitions we had in the confJson. To work
around this cascadeDeleteJsonFile and loadCascadeDelete can be set to save
and/or load hte cascade delete configuration separately.

So now, we have three things we can configure in properties to generate and
load:
* confFile, loadConf
* metadataFile, loadMetadata,
* cascadeDeleteFile, loadCascadeDelete