Allow setting null/empty value for metadataJsonFile and schemaFile
By default metadata.json and hasura-json-schema.json are set for
metadataJsonFile and schemaFile when HasuraConfigurator initialized as a bean.
However it is impossible to set null values for these form .properties files.
To supportt this the value "null" and empty strings will be considered as
null values and in this case metadata and schema file will not be written.