Add automatic relationship generation for fields referring to @Entity
Besides explicitly defining all data of a @HasuraRelationship if a field
refers to a @Entity then HasuraActionGenerator autoamtically defines the
relationship, and modifies the graphql type of the containing entity to
include a fileNameId field where the action needs to provide  the ID to the
entity that is related.

Also added testActionGeneration and proper test in HasuraActionGeneratorTest.
Need to add more thorough tests for edge case, etc.