Fix handling id fields for @HasuraPermission.excludeFields
The object own ID field doesn't come form classMetadata.propertyNames
but from classMetadata.identifierPropertyName so this has to be added
to the fields we use for calculation.

On the other hand we should ignore properties that do not have an actual
column in the table. ie. a remote entity mapped relationship.