Add support for @HasuraPermissions on @ManyToMany fields
The entities defines for @ManyToMany joins are not exposed as Java objects
by Hibernate, so we cannot use the existing permission processing algorithms,
but had to add a specific one based on M2MData instead of EntityType.
The two processing methods still use many functions in common, so the logics
are pretty much the same except the source of information (M2MData vs
EntityType).