Allow overriding output type via @HasuraAction
By default we take the return type of action method as the output type, but
now this can be overriden using @HasuraAction.outputType. This can be useful
in cases when the actual method return type is different from the one
which needs to exposed to graphql. Eg. in case of Spring when we use
ResponseEntity as the return type and the actual type parameter is the output
type we want to use.