Add support for class based @HasuraEnum-s
Instead of defining an @Entity as enum, you can now also define it as
a normal class. In this case the hasura enum values will be taken from
the filan static string fields annotated using @HasuraEnumValue.

The reason for using normal classes instead of enums could be that when using
these classes in a Hibarnate based Spring app you need to provide a public
no-arg constructor for all @Entity classes, however for enums a default private
one is already defined and a no-arg cosntructor cannot be defined.

If hasuraconf is used only to define the hasura metadata then using enums
is preferred.