Refactor to work with Hasura Metadata V3
Instead of generating JSON as we process JSP/Hasura annotations, we build
a HasuraMetadataV3 model. HasuraConfigurator#configure() now returns a
HasuraConfiguration object, which can be used to generate the metadata JSON,
the SQL-s to configure enums, functions, etc. and also the JsonSchema.
HasuraConfigurator doesn't generate and load these JSONs, however, but the
HasuraConfiguration#toMetadataJSON(), HasuraConfiguration#toBulkRunSql() and
HasuraConfigurator#load...() funtions can be used fro this.

From the original tests now the HasuraActionGeneratorTest and
PermissionAnnotationProcessorTest pass, CustomRootFieldNameProviderTest
HasuraActionGeneratorTest, HasuraConfiguratorIntegrationTests needs to
be updated to match the new HasuraConfigurator API.