Add separate fields for configs with run_slq-s
Instead of one runSqls array, we have separate configs for enums, computed
fields and cascaded delete fields so we can always check the source of these
sqls. The utility extension fields HasuraConfiguration.runSqls and
HasuraConfiguration.kt.bulkRunSqlJson provide access to the merged run_sql-s
as a list of JsobObjects or as a bulk of run_sql schema operation.