Add tests for native Spring actions via for HasuraActionController
SignInAction implements anaction with a request transform, which forwards the
action call to /api/auth/signin provided by the AuthController REST controller.

StartTaskAction is a @RestController operation, which implement the startTask
action proxied via HasuraActionController and the body is transformed from
Hasura action payload to the actual request format of the startTask action
using HasuraActionFilter and ActionRequestWrapper.